[PAL,LibOS,common] Add file recovery support for encrypted files
Previously, a fatal error during writes to encrypted files could cause
file corruption due to incorrect GMACs and/or encryption keys.

To address this, we introduce a file recovery mechanism using a "shadow"
recovery file that stores data about to change and an update flag in the
metadata node indicating the start of a write transaction. During file
flush, all cached blocks that are about to change are saved to the
recovery file in the format of physical node numbers (offsets) plus
encrypted block data. Before saving the main file contents, the
`update_flag` is set in the file's metadata node and cleared only when
the transaction is complete. If an encrypted file is opened and the
`update_flag` is set, a recovery process starts to revert partial
changes using the recovery file, returning to the last known good state.
The "shadow" recovery file is cleaned up on file close.

This commit adds a new mount parameter `enable_recovery = [true|false]`
for encrypted files mounts to optionally enable this feature. We extend
the file flush logic of protected files (pf) to include the recovery
file dump and the setting/unsetting of the update flag. We make changes
to the public pf APIs: the `pf_open()` API is extended to make the pf
aware of the underlying recovery file managed by the LibOS, and recovery
information (e.g., whether the pf needs recovery) is exposed back to the
LibOS via a new `pf_get_recovery_info()` API. To facilitate the LibOS to
initiate a file recovery process on file open, a new PAL API
`PalEncryptedFileRecovery()` is introduced.

Signed-off-by: Kailun Qin <kailun.qin@intel.com>